refactoring - Moving job-related features to `jobs`
