+++ features/clock/api.ts
@@ -7,7 +7,7 @@ import { LocationInput, Employers } from '../../types';
 import * as FileSystem from 'expo-file-system';
 import { useQuery } from 'react-query';
 import moment from 'moment';
-import { useNumJobsNeedEntryThisWeek } from '../job/api';
+import { useNumJobsNeedEntryThisWeek } from '../jobs/api';
 
 export const fetchActiveShift = async () => {
     const client = getClient(store);
